move ProductOptions class to separate cs file
