+++ refactor-me/Models/ProductOption.cs
@@ -1,0 +1,63 @@
+using System;
+using System.Data.SqlClient;
+using Newtonsoft.Json;
+
+namespace refactor_me.Models
+{
+    public class ProductOption
+    {
+        public Guid Id { get; set; }
+
+        public Guid ProductId { get; set; }
+
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+
+        [JsonIgnore]
+        public bool IsNew { get; }
+
+        public ProductOption()
+        {
+            Id = Guid.NewGuid();
+            IsNew = true;
+        }
+
+        public ProductOption(Guid id)
+        {
+            IsNew = true;
+            var conn = Helpers.NewConnection();
+            var cmd = new SqlCommand($"select * from productoption where id = '{id}'", conn);
+            conn.Open();
+
+            var rdr = cmd.ExecuteReader();
+            if (!rdr.Read())
+                return;
+
+            IsNew = false;
+            Id = Guid.Parse(rdr["Id"].ToString());
+            ProductId = Guid.Parse(rdr["ProductId"].ToString());
+            Name = rdr["Name"].ToString();
+            Description = (DBNull.Value == rdr["Description"]) ? null : rdr["Description"].ToString();
+        }
+
+        public void Save()
+        {
+            var conn = Helpers.NewConnection();
+            var cmd = IsNew ?
+                new SqlCommand($"insert into productoption (id, productid, name, description) values ('{Id}', '{ProductId}', '{Name}', '{Description}')", conn) :
+                new SqlCommand($"update productoption set name = '{Name}', description = '{Description}' where id = '{Id}'", conn);
+
+            conn.Open();
+            cmd.ExecuteNonQuery();
+        }
+
+        public void Delete()
+        {
+            var conn = Helpers.NewConnection();
+            conn.Open();
+            var cmd = new SqlCommand($"delete from productoption where id = '{Id}'", conn);
+            cmd.ExecuteReader();
+        }
+    }
+}
